--- src/redux/actionTypes/profileActionTypes.js
@@ -1,1 +1,0 @@
-export const VIEW_USERS_INFO = "VIEW_USERS_INFO";
