+++ src/redux/actionTypes/userActionTypes.js
@@ -1,0 +1,1 @@
+export const VIEW_USERS = "VIEW_USERS";
