+++ src/redux/reducers/userReducers/viewUserReducer.js
@@ -1,13 +1,13 @@
-import { VIEW_USERS_INFO } from "../../actionTypes/profileActionTypes";
+import { VIEW_USERS } from "../../actionTypes/userActionTypes";
 
 import initialState from "../../store/initialState";
 
 const viewUsersReducer = (state = initialState, action) => {
   switch (action.type) {
-    case VIEW_USERS_INFO:
+    case VIEW_USERS:
       return {
         ...state,
-        viewProfileInfo: action.payload,
+        userInfo: action.payload,
       };
 
     default:
